bad_elf_symbols*.pl: Handle unique global symbols
According to nm(1), symbols of type "u" are unique global symbols.
This is a GNU extension to the standard set of ELF symbol bindings. For
such a symbol the dynamic linker will make sure that in the entire
process there is just one symbol with this name and type in use.