rpmelfsym.xs: make writes atomic by flushing at PIPE_BUF
I understand that write(2) with up to PIPE_BUF bytes is guaranteed
to be atomic, if either 1) writing to a pipe or fifo; or 2) writing
to a regular file opened in O_APPEND mode.  The latter is probably
not specified by SUS, but it is obviously the only possibility to
implement log files written by multiple processes.

So now we can have e.g. two threads which either pipe to sort(1) or
append to a regular file without interleaving the lines.