rpmelfsym.pm: implemented parallel collect_bad_elfsym
AMD has announced new 16-core CPUs, which will be used in Supermicro's
1U dual-socket servers (that's 32 cores per 1U).  Clearly, something must
be done in response to this new challenge. :)

So here I introduce new design with two parallel "workers" which execute
collect_bad_elfsym code and append output to ref/def/seq (using atomic
write techinque implemented in a previous commit).

My initial design was even more sophisticated - it also involved piping
to sort(1):

        collect1                collect2
         |    |                  |    |
         | /---------------------/    |
         | |  \---------------------\ |
         | |                        | |
         v v                        v v
        sort ref                sort def

However, when reading from a pipe, it seems that sort(1) does not
run in parallel mode (apparently due conservative "pipe size"
heuristics).  The result is that both workers are stalled due to
slow single-threaded sorting.  So, however appealing this new design
might be, it is best to use temporary files for now.

Note that parallel workers are especially useful when the cache
is populated for the first time, because expensive lzma decompression
and cpio extraction routines will run in parallel about twice as fast.