bad_elf_symbols_dircmp.pl: optimize def0 usage for parallel join
Note that def0 (symbols defined by rpms common for both DIR1 and DIR2)
is biggest file in the whole process (430M unsorted, 250M sorted).
We need to join it against ref0, ref1, and ref2.  Also note that that
ref0 is the second most biggest file (150M).  So we have 3 joins
against the big file, and the first join is considerably more
expensive because both its files are big.  To me, it's a perfect
opportunity to run the first join in background.

Some benchmarking: on armor.altlinux (ancient Xeon 5110, 4 cores)
the shell part of the script took 32s before the optimization,
and now it takes only 27s.  That's 15% for just reordering a few
commands and putting an ampersand in the right place.