rpmelfsym.pm: revamped temporary file handling, process nm(1) stderr
File::Temp is no longer required.  Only two anonymous files are created
on the first iteration: one for ELF content, and the other for nm(1) stderr.
This works, but depends on peculiarities of /proc/$$/fd implementaiton:
i.e. whether it has open(2) or dup(2) semantics, complicated by the fact
that $$ might not be "self".  So now we need to remember to truncate
files at the end of each iteration, and to seek to the beginning before
each access.

nm(1) stderr output is now correctly identified.  I.e, it used to be:

nm: /tmp/.private/at/k9NG_GcnWa: no symbols

And now it is:

haspd-7.40-alt4.0.M70P.1.i586.rpm: nm: /usr/sbin/hasplmd: no symbols